Refactor Items
This change refactors items from a sum type into a nested interface.
This makes items a lot easier to deal with (since we don't need to check
the item type all the time), and it allows items to affect different
stats (so a drill might be able to give you some boosters, etc).